Replaced inner class with HttpAsyncTask.java
