+++ app/src/main/java/hr/fer/android/sglab/qr/InterfaceActivity.java
@@ -1,40 +1,24 @@
 package hr.fer.android.sglab.qr;
 
-import android.annotation.SuppressLint;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Base64;
 import android.view.animation.Animation;
 import android.view.animation.LinearInterpolator;
 import android.view.animation.RotateAnimation;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import java.io.BufferedReader;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-
 public class InterfaceActivity extends AppCompatActivity {
-
-    private static final String DEFAULT_ENDPOINT = "http://10.0.2.2:8080/api/sglab/machines/";
-    private static final String USERNAME = "sglabadmin";
-    private static final String PASSWORD = "sglabadmin";
-
-    private static String encoded;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
         startLoading();
-
-        encoded = Base64.encodeToString((USERNAME + ":" + PASSWORD).getBytes(), Base64.NO_WRAP).trim();
-
         Toast.makeText(this, "Scanned:" + getIntent().getStringExtra("QRResult"), Toast.LENGTH_LONG).show();
 
-        new httpGet().execute(getIntent().getStringExtra("QRResult"));
+        new HttpAsyncTask(this).execute(getIntent().getStringExtra("QRResult"));
 
     }
 
@@ -60,36 +44,7 @@ public class InterfaceActivity extends AppCompatActivity {
         findViewById(R.id.logo_circle3).startAnimation(anim3);
     }
 
-    private class httpGet extends AsyncTask<String, Integer, String> {
-        protected String doInBackground(String... urls) {
-            try {
-                final URL url = new URL(DEFAULT_ENDPOINT +  getIntent().getStringExtra("QRResult"));
-                final HttpURLConnection con = (HttpURLConnection) url.openConnection();
-
-                //When
-                con.setRequestProperty("Authorization", "Basic " + encoded);
-                con.setRequestProperty("Content-Type", "application/json");
-                con.setRequestMethod("GET");
-
-
-                StringBuilder result = new StringBuilder();
-                BufferedReader rd = new BufferedReader(new InputStreamReader(con.getInputStream()));
-                String line;
-                while ((line = rd.readLine()) != null) {
-                    result.append(line);
-                }
-                rd.close();
-                return result.toString();
-            } catch (Exception ignored) {
-                return null;
-            }
-        }
-
-        protected void onPostExecute(String result) {
-            processData(result);
-        }
-    }
-
+    //Called when data is received by the HttpAsyncTask
     protected void processData(String data) {
         if (data != null) {
             setContentView(R.layout.hydro_interface);
@@ -99,5 +54,4 @@ public class InterfaceActivity extends AppCompatActivity {
             Toast.makeText(this, "Failed...", Toast.LENGTH_LONG).show();
         }
     }
-
 }
